+++ src/lib/components/display/datestamp/DateStamp.svelte
@@ -18,8 +18,8 @@
         actions?: IForwardedActions;
         element?: HTMLTimeElement;
 
-        calendar: string;
-        locale: string;
+        calendar?: string;
+        locale?: string;
 
         day?: Intl.DateTimeFormatOptions["day"];
         month?: Intl.DateTimeFormatOptions["month"];
@@ -37,8 +37,8 @@
     let _class = "";
     export {_class as class};
 
-    export let calendar: $$Props["calendar"] = DEFAULT_CALENDAR;
-    export let locale: $$Props["locale"] = DEFAULT_LOCALE;
+    export let calendar: $$Props["calendar"] = undefined;
+    export let locale: $$Props["locale"] = undefined;
 
     export let day: $$Props["day"] = undefined;
     export let month: $$Props["month"] = undefined;
@@ -75,8 +75,8 @@
     on:pointerout
     on:pointerup
 >
-    {_date.toLocaleString(locale, {
-        calendar,
+    {_date.toLocaleString(locale ?? DEFAULT_LOCALE, {
+        calendar: calendar ?? DEFAULT_CALENDAR,
         day: _options.day,
         month: _options.month,
         weekday: _options.weekday,