+++ src/lib/components/display/datetimestamp/DateTimeStamp.svelte
@@ -19,8 +19,8 @@
         actions?: IForwardedActions;
         element?: HTMLTimeElement;
 
-        calendar: string;
-        locale: string;
+        calendar?: string;
+        locale?: string;
 
         day?: Intl.DateTimeFormatOptions["day"];
         month?: Intl.DateTimeFormatOptions["month"];
@@ -43,8 +43,8 @@
     let _class = "";
     export {_class as class};
 
-    export let calendar: $$Props["calendar"] = DEFAULT_CALENDAR;
-    export let locale: $$Props["locale"] = DEFAULT_LOCALE;
+    export let calendar: $$Props["calendar"] = undefined;
+    export let locale: $$Props["locale"] = undefined;
 
     export let day: $$Props["day"] = undefined;
     export let month: $$Props["month"] = undefined;
@@ -91,8 +91,8 @@
     on:pointerout
     on:pointerup
 >
-    {_datetime.toLocaleString(locale, {
-        calendar,
+    {_datetime.toLocaleString(locale ?? DEFAULT_LOCALE, {
+        calendar: calendar ?? DEFAULT_CALENDAR,
         day: _options.day,
         month: _options.month,
         weekday: _options.weekday,