+++ src/lib/components/overlays/offscreen/Offscreen.svelte
@@ -33,6 +33,7 @@
         captive?: boolean;
         dismissible?: boolean;
         logic_id?: string;
+        once?: boolean;
         state?: boolean;
 
         placement?: PROPERTY_PLACEMENT;
@@ -58,6 +59,7 @@
     export let captive: $$Props["captive"] = undefined;
     export let dismissible: $$Props["dismissible"] = undefined;
     export let logic_id: $$Props["logic_id"] = "";
+    export let once: $$Props["once"] = undefined;
     export let state: $$Props["state"] = undefined;
 
     export let placement: $$Props["placement"] = undefined;
@@ -71,8 +73,17 @@
 
     let _previous_state = state;
 
-    function on_change(event: Event) {
+    function on_change(event: Event): void {
         state = (event.target as HTMLInputElement).checked;
+    }
+
+    function on_click_inside(event: MouseEvent): void {
+        if (once) {
+            const target = event.target as HTMLElement;
+            if (target.tagName !== "LABEL" && target.getAttribute("for") !== logic_id) {
+                state = false;
+            }
+        }
     }
 
     $: $_state = state as boolean;
@@ -111,6 +122,7 @@
         placement,
     })}
     use:forward_actions={{actions}}
+    on:click={on_click_inside}
     on:click
     on:contextmenu
     on:dblclick