+++ src/lib/components/overlays/overlay/Overlay.svelte
@@ -34,6 +34,7 @@
         captive?: boolean;
         dismissible?: boolean;
         logic_id?: string;
+        once?: boolean;
         state?: boolean;
 
         orientation?: PROPERTY_ORIENTATION_Y_BREAKPOINT;
@@ -64,6 +65,7 @@
     export let captive: $$Props["captive"] = undefined;
     export let dismissible: $$Props["dismissible"] = undefined;
     export let logic_id: $$Props["logic_id"] = "";
+    export let once: $$Props["once"] = undefined;
     export let state: $$Props["state"] = undefined;
 
     export let orientation: $$Props["orientation"] = undefined;
@@ -81,8 +83,17 @@
 
     let _previous_state = state;
 
-    function on_change(event: Event) {
+    function on_change(event: Event): void {
         state = (event.target as HTMLInputElement).checked;
+    }
+
+    function on_click_inside(event: MouseEvent): void {
+        if (once) {
+            const target = event.target as HTMLElement;
+            if (target.tagName !== "LABEL" && target.getAttribute("for") !== logic_id) {
+                state = false;
+            }
+        }
     }
 
     $: $_state = state as boolean;
@@ -124,6 +135,7 @@
         "spacing-y": spacing_y,
     })}
     use:forward_actions={{actions}}
+    on:click={on_click_inside}
     on:click
     on:contextmenu
     on:dblclick